fix: fix switch form
